requirements.txt
absl-py==0.8.1
altgraph==0.17
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
Cython==0.29.14
dj-database-url==0.5.0
Django==3.0.5
django-bootstrap4==1.1.1
django-ckeditor==5.9.0
django-heroku==0.3.1
django-js-asset==1.2.2
future==0.18.2
gast==0.3.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.17.4
pandas==0.25.3
patsy==0.5.1
pefile==2019.4.18
Pillow==7.1.1
psycopg2==2.8.5
pygeoip==0.3.2
PyInstaller==3.6
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
pywin32-ctypes==0.2.0
PyYAML==5.2
requests==2.23.0
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
soupsieve==2.0
sqlparse==0.3.1
statsmodels==0.10.2
urllib3==1.25.9
whitenoise==5.0.1
